% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-id.R
\name{extract_id}
\alias{extract_id}
\title{Extract an identifier using regex}
\usage{
extract_id(input, pattern, group)
}
\arguments{
\item{input}{The string to search}

\item{pattern}{The regular expression to match}

\item{group}{The group within the regular expression}
}
\value{
The matching identifier from the string
}
\description{
This is an internal helper function for exported functions such as
extract_labno, extract_worksheet and extract_suffix to reduce code
duplication when checking inputs and outputs.
}
