% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-igene-rno.R
\name{extract_igene_rno}
\alias{extract_igene_rno}
\title{Extract iGene R number from a character string}
\usage{
extract_igene_rno(input)
}
\arguments{
\item{input}{A character string containing an iGene R number}
}
\value{
The iGene R number as a string
}
\description{
The iGene database uses R numbers to identify individual referrals.
}
\details{
The R number is R, followed by the year as 2 digits (2026 = 26) and then
a random string of 5 alphanumeric characters.
}
\examples{
extract_igene_rno("WS123456_R26-12AB")
}
