% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-labno.R
\name{extract_labno}
\alias{extract_labno}
\title{Extract lab number from a string}
\usage{
extract_labno(input)
}
\arguments{
\item{input}{A character string containing a lab number}
}
\value{
The lab number as a string
}
\description{
The DNA "lab number" (labno) is an eight digit numeric string which
is used to identify DNA samples at the Manchester Genomic Laboratory
Hub.
}
\details{
The labno has the year as the first 2 digits (2026 = 26) and the remaining
6 digits increase from 1.
Example: the first sample of 2026 would have labno 26000001.
}
\examples{
extract_labno("WS123456_12345678")
}
