% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-suffix.R
\name{extract_suffix}
\alias{extract_suffix}
\title{Extract a sample replicate suffix from a string}
\usage{
extract_suffix(input)
}
\arguments{
\item{input}{A character string containing a lab number and a suffix}
}
\value{
A character string of the suffix. Inputs without a suffix with
return "".
}
\description{
When multiple replicates of a DNA lab number (labno) are tested on the same
worksheet, a letter is added directly after each labno.
}
\details{
Example: 12345678a, 12345678b, 12345678c
}
\examples{
extract_suffix("WS123456_12345678a_replicate1")
}
