% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-clc-msi-locus-tables.R
\name{join_clc_msi_locus_tables}
\alias{join_clc_msi_locus_tables}
\title{Join information in CLC MSI Report locus section}
\usage{
join_clc_msi_locus_tables(
  df,
  row_distance_locus_to_table = 2,
  row_distance_stability_to_table = 1
)
}
\arguments{
\item{df}{The output from mutate_clc_msi_row_numbers}

\item{row_distance_locus_to_table}{The number of rows between the locus name
and the table of length information for that locus}

\item{row_distance_stability_to_table}{The number of rows between the
stability prediction for a locus and the table of length information for that
locus.}
}
\value{
A list of tables with the first item being the final joined table
}
\description{
Join information in CLC MSI Report locus section
}
