% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-filename.R
\name{mutate_filename}
\alias{mutate_filename}
\title{Add filename and filepath to a dataframe}
\usage{
mutate_filename(df, filepath)
}
\arguments{
\item{df}{A dataframe}

\item{filepath}{The filepath of the file the dataframe has been read from}
}
\value{
The dataframe with filepath and filename columns
}
\description{
This function is a helper function to reduce code duplication for
the read_.*_with_filename series of functions.
}
