% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-ids.R
\name{mutate_ids}
\alias{mutate_ids}
\title{Add sample identifiers from filename as additional data columns}
\usage{
mutate_ids(df, id_col = filename)
}
\arguments{
\item{df}{The dataframe to add identifiers to}

\item{id_col}{The column in the dataframe which includes the filename}
}
\value{
The original dataframe with additional columns of identifiers
from the filename column
}
\description{
This function extracts the 3 common identifiers (worksheet, labno and
suffix) which are included in filenames.
}
\details{
Patient name and NGS panel are both not included in \code{mutate_ids}, as they
are not always present in filenames.
}
