% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-clc-msi-locus-section.R
\name{read_clc_msi_locus_section}
\alias{read_clc_msi_locus_section}
\title{Read the section of a CLC MSI Report Excel which contains locus lengths}
\usage{
read_clc_msi_locus_section(
  filepath,
  locus_section_title = "Loci Length Distributions",
  ...
)
}
\arguments{
\item{filepath}{The filepath to the CLC MSI Report}

\item{locus_section_title}{The string which marks the the change from the
summary section of the report to the individual locus length tables.}

\item{...}{Additional arguments to pass to read_excel if needed}
}
\value{
The section of the Excel sheet containing the locus length tables
with every column formatted as text.
}
\description{
Read the section of a CLC MSI Report Excel which contains locus lengths
}
