% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_ids.R
\name{read_csv_ids}
\alias{read_csv_ids}
\title{Read a csv file and add identifiers}
\usage{
read_csv_ids(filepath)
}
\arguments{
\item{filepath}{Filepath to a csv file with worksheet, labno and suffix
identifiers in the filename}
}
\value{
The csv file as a dataframe with additional columns containing the
identifiers from the filename
}
\description{
Read a csv file and add identifiers
}
