% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-csv-with-filename.R
\name{read_csv_with_filename}
\alias{read_csv_with_filename}
\title{Read a csv file and add the filename as a column}
\usage{
read_csv_with_filename(filepath)
}
\arguments{
\item{filepath}{Filepath to a csv file.}
}
\value{
The csv file as a dataframe with the filename and filepath as
additional columns
}
\description{
Read a csv file and add the filename as a column
}
