% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-excel-with-filename.R
\name{read_excel_with_filename}
\alias{read_excel_with_filename}
\title{Read an Excel file with additional columns for filename}
\usage{
read_excel_with_filename(filepath, ...)
}
\arguments{
\item{filepath}{The filepath to the Excel file}

\item{...}{Extra arguments to pass to read_excel (range, skip etc)}
}
\value{
The file as a dataframe, with additional filepath and filename
columns
}
\description{
This is a helper function which can be combined with mutate_ids() when
reading in files.
}
