% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-json-as-df.R
\name{read_json_as_df}
\alias{read_json_as_df}
\title{Read a table from a JSON file as a dataframe}
\usage{
read_json_as_df(
  filepath,
  json_list = function(x) x$data$summary$quality_control$table_1
)
}
\arguments{
\item{filepath}{The filepath for the JSON file}

\item{json_list}{A function which includes the list structure to the table of
interest within the JSON file. The default argument is for the table
containing quality control information in a PanSolid JSON file.}
}
\value{
The table as a dataframe, including filename and filepath columns
}
\description{
Read a table from a JSON file as a dataframe
}
