% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-vcf-with-filename.R
\name{read_vcf_with_filename}
\alias{read_vcf_with_filename}
\title{Read a VCF file into R as a data frame}
\usage{
read_vcf_with_filename(filepath, rename_last_col = TRUE)
}
\arguments{
\item{filepath}{The full filepath to a VCF}

\item{rename_last_col}{TRUE or FALSE, whether to rename the final column
of the gt table, which often includes a file specific heading. Renaming
the final column is helpful for reading a collating multiple VCFs from
different samples.}
}
\value{
A dataframe of the fix and gt tables from the VCF
}
\description{
Read a VCF file into R as a data frame
}
