% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex-ids.R
\name{regex_ids}
\alias{regex_ids}
\title{Regular expressions for sample identifiers}
\usage{
regex_ids()
}
\value{
A named list of regular expressions for different identifiers
}
\description{
Labno is specified as having non-digits on either side, to prevent
extracting 8 digits from a longer string of digits.
}
\examples{
regex_ids()$labno_suffix
}
